refactor: Exception 합치기
