+++ src/main/java/com/travel/role/domain/comment/service/CommentService.java
@@ -14,15 +14,15 @@ import com.travel.role.domain.comment.dto.CommentListResDTO;
 import com.travel.role.domain.comment.dto.CommentReqDTO;
 import com.travel.role.domain.comment.dto.CommentResDTO;
 import com.travel.role.domain.comment.entity.Comment;
-import com.travel.role.domain.comment.exception.CommentInfoNotFoundException;
+import com.travel.role.global.exception.comment.CommentInfoNotFoundException;
 import com.travel.role.domain.room.repository.RoomParticipantRepository;
 import com.travel.role.domain.room.repository.RoomRepository;
 import com.travel.role.domain.room.entity.Room;
 import com.travel.role.domain.user.repository.UserRepository;
 import com.travel.role.domain.user.entity.User;
-import com.travel.role.domain.user.exception.RoomInfoNotFoundException;
-import com.travel.role.domain.user.exception.UserInfoNotFoundException;
-import com.travel.role.domain.user.exception.UserNotParticipateRoomException;
+import com.travel.role.global.exception.user.RoomInfoNotFoundException;
+import com.travel.role.global.exception.user.UserInfoNotFoundException;
+import com.travel.role.global.exception.user.UserNotParticipateRoomException;
 import com.travel.role.global.exception.ResourceOperationAccessDeniedException;
 
 import lombok.RequiredArgsConstructor;
