+++ src/main/java/com/travel/role/domain/room/service/RoomService.java
@@ -26,14 +26,14 @@ import com.travel.role.domain.room.dto.InviteResponseDTO;
 import com.travel.role.domain.room.dto.MakeRoomRequestDTO;
 import com.travel.role.domain.room.dto.MemberDTO;
 import com.travel.role.domain.room.dto.RoomResponseDTO;
-import com.travel.role.domain.room.exception.AlreadyExistInRoomException;
-import com.travel.role.domain.room.exception.InvalidInviteCode;
-import com.travel.role.domain.room.exception.InvalidLocalDateException;
-import com.travel.role.domain.room.exception.UserHaveNotPrivilegeException;
+import com.travel.role.global.exception.room.AlreadyExistInRoomException;
+import com.travel.role.global.exception.room.InvalidInviteCode;
+import com.travel.role.global.exception.room.InvalidLocalDateException;
+import com.travel.role.global.exception.room.UserHaveNotPrivilegeException;
 import com.travel.role.domain.user.repository.UserRepository;
 import com.travel.role.domain.user.entity.User;
-import com.travel.role.domain.user.exception.RoomInfoNotFoundException;
-import com.travel.role.domain.user.exception.UserInfoNotFoundException;
+import com.travel.role.global.exception.user.RoomInfoNotFoundException;
+import com.travel.role.global.exception.user.UserInfoNotFoundException;
 import com.travel.role.global.auth.token.UserPrincipal;
 import com.travel.role.global.util.PasswordGenerator;
