+++ src/main/java/com/travel/role/domain/room/service/TravelEssentialService.java
@@ -18,9 +18,9 @@ import com.travel.role.domain.room.dto.TravelEssentialReqDTO;
 import com.travel.role.domain.room.dto.TravelEssentialResDTO;
 import com.travel.role.domain.user.repository.UserRepository;
 import com.travel.role.domain.user.entity.User;
-import com.travel.role.domain.user.exception.RoomInfoNotFoundException;
-import com.travel.role.domain.user.exception.UserInfoNotFoundException;
-import com.travel.role.domain.user.exception.UserNotParticipateRoomException;
+import com.travel.role.global.exception.user.RoomInfoNotFoundException;
+import com.travel.role.global.exception.user.UserInfoNotFoundException;
+import com.travel.role.global.exception.user.UserNotParticipateRoomException;
 
 import lombok.RequiredArgsConstructor;
