+++ src/main/java/com/travel/role/global/auth/service/CustomOAuth2UserService.java
@@ -18,7 +18,7 @@ import com.travel.role.domain.user.entity.User;
 import com.travel.role.global.auth.oauth.OAuth2UserInfo;
 import com.travel.role.global.auth.oauth.OAuthAttributes;
 import com.travel.role.global.auth.token.UserPrincipal;
-import com.travel.role.domain.user.exception.AlreadyExistUserException;
+import com.travel.role.global.exception.user.AlreadyExistUserException;
 
 import lombok.RequiredArgsConstructor;
