+++ src/main/java/com/travel/role/global/exception/room/InvalidInviteCode.java
@@ -1,4 +1,4 @@
-package com.travel.role.domain.room.exception;
+package com.travel.role.global.exception.room;
 
 import static com.travel.role.global.exception.ExceptionMessage.*;
