+++ src/test/java/com/travel/role/domain/room/service/RoomServiceTest.java
@@ -19,15 +19,15 @@ import com.travel.role.domain.room.repository.ParticipantRoleRepository;
 import com.travel.role.domain.room.repository.RoomRepository;
 import com.travel.role.domain.room.entity.Room;
 import com.travel.role.domain.room.dto.MakeRoomRequestDTO;
-import com.travel.role.domain.room.exception.AlreadyExistInRoomException;
-import com.travel.role.domain.room.exception.InvalidInviteCode;
-import com.travel.role.domain.room.exception.InvalidLocalDateException;
-import com.travel.role.domain.room.exception.UserHaveNotPrivilegeException;
+import com.travel.role.global.exception.room.AlreadyExistInRoomException;
+import com.travel.role.global.exception.room.InvalidInviteCode;
+import com.travel.role.global.exception.room.InvalidLocalDateException;
+import com.travel.role.global.exception.room.UserHaveNotPrivilegeException;
 import com.travel.role.domain.user.repository.UserRepository;
 import com.travel.role.domain.user.entity.Provider;
 import com.travel.role.domain.user.entity.Role;
 import com.travel.role.domain.user.entity.User;
-import com.travel.role.domain.user.exception.UserInfoNotFoundException;
+import com.travel.role.global.exception.user.UserInfoNotFoundException;
 import com.travel.role.global.auth.token.UserPrincipal;
 import com.travel.role.global.exception.ExceptionMessage;
 import com.travel.role.global.util.PasswordGenerator;
